requirements.txt
anthropic
python-dotenv
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-docx
